Kotlin cleanup and consolidation
