--- android/src/main/java/com/willowtreeapps/signinwithapplebutton/SignInWithAppleCallback.kt
@@ -1,19 +1,0 @@
-package com.willowtreeapps.signinwithapplebutton
-
-interface SignInWithAppleCallback {
-
-    fun onSignInWithAppleSuccess(code: String, id_token: String, state: String, user: String)
-
-    fun onSignInWithAppleFailure(error: Throwable)
-
-    fun onSignInWithAppleCancel()
-}
-
-internal fun SignInWithAppleCallback.toFunction(): (SignInWithAppleResult) -> Unit =
-    { result ->
-        when (result) {
-            is SignInWithAppleResult.Success -> onSignInWithAppleSuccess(result.code,result.id_token,result.state,result.user)
-            is SignInWithAppleResult.Failure -> onSignInWithAppleFailure(result.error)
-            is SignInWithAppleResult.Cancel -> onSignInWithAppleCancel()
-        }
-    }
