--- android/src/main/java/com/willowtreeapps/signinwithapplebutton/SignInWithAppleResult.kt
@@ -1,9 +1,0 @@
-package com.willowtreeapps.signinwithapplebutton
-
-sealed class SignInWithAppleResult {
-    data class Success(val code: String, val id_token: String, val state: String, val user: String) : SignInWithAppleResult()
-
-    data class Failure(val error: Throwable) : SignInWithAppleResult()
-
-    object Cancel : SignInWithAppleResult()
-}
